INT-2825 Remove Temporary 2.2. Interface
AbstractTransactionSynchronizingPollingEndpoint was introduced
late in the 2.2. release process to avoid a breaking change to
AbstractPollingEndpoint.

For 3.0, its methods are now pulled-up into APE.
