+++ spring-integration-core/src/main/java/org/springframework/integration/endpoint/SourcePollingChannelAdapter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2012 the original author or authors.
+ * Copyright 2002-2013 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ import org.springframework.util.Assert;
  * @author Oleg Zhurakousky
  * @author Gary Russell
  */
-public class SourcePollingChannelAdapter extends AbstractTransactionSynchronizingPollingEndpoint
+public class SourcePollingChannelAdapter extends AbstractPollingEndpoint
 		implements TrackableComponent {
 
 	private volatile MessageSource<?> source;
